requirements.txt
# Core
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.2.0

# Celery + Redis
celery==5.4.0
redis==5.0.8

# Stripe
stripe==10.8.0

# API Docs
drf-spectacular==0.27.2

# Static files / Production server
whitenoise==6.7.0
gunicorn==22.0.0

# Environment
python-dotenv==1.0.1
